Add bump mapping feature and update HUD controls
- Introduced bump mapping toggle functionality in input handling.
- Updated app_state.h to include bump mapping state.
- Adjusted draw_hud to display bump mapping toggle instructions.
- Refactored bump mapping condition in draw_sphere_lod to use dynamic variable.
- Corrected variable type for file size in read_file function.

+++ src/utils.c
@@ -21,7 +21,7 @@ char* read_file(const char* path) {
 
     // encontra final
     fseek(f, 0, SEEK_END);
-    long sz = ftell(f);
+    size_t sz = ftell(f);
     rewind(f);
 
     // aloca espaço pro buffer (+1 para o \0)
